Reorganizowano częściowo pliki - problemy z budową obiektów uniemożliwiają kompilację
