+++ Engine/Engine.vcxproj
@@ -140,7 +140,7 @@
       <LanguageStandard_C>stdc17</LanguageStandard_C>
       <PrecompiledHeader>Create</PrecompiledHeader>
       <PrecompiledHeaderFile>Engine.h</PrecompiledHeaderFile>
-      <AdditionalIncludeDirectories>C:\Users\tymek\source\repos\Game_engine_1\Engine\Level_control;C:\Users\tymek\source\repos\Game_engine_1\Engine\Core;C:\Users\tymek\source\repos\Game_engine_1\Engine\resources;C:\Users\tymek\source\repos\Game_engine_1\Engine\pch;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>C:\Users\tymek\source\repos\Game_engine_1\Engine\Core;C:\Users\tymek\source\repos\Game_engine_1\Engine\Engine_pch;C:\Users\tymek\source\repos\Game_engine_1\Engine\Level_control;C:\Users\tymek\source\repos\Game_engine_1\Engine\resources;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
     </ClCompile>
     <Link>
       <SubSystem>Console</SubSystem>
@@ -151,20 +151,18 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="Core\CoreDefinitions.cpp" />
     <ClCompile Include="Core\CoreMinimal.cpp">
       <PrecompiledHeaderFile Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Engine.h</PrecompiledHeaderFile>
+      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="Level_control\GameController.cpp" />
-    <ClCompile Include="pch\Engine.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\resources\resource.h" />
-    <ClInclude Include="Core\CoreDefinitions.h" />
     <ClInclude Include="Core\CoreMinimal.h" />
+    <ClInclude Include="Engine_pch\Engine.h" />
     <ClInclude Include="Level_control\GameController.h" />
     <ClInclude Include="Level_control\GameLevel.h" />
-    <ClInclude Include="pch\Engine.h" />
     <ClInclude Include="resource.h" />
   </ItemGroup>
   <ItemGroup>
