+++ Engine/Engine.vcxproj.filters
@@ -13,9 +13,6 @@
       <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
       <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
     </Filter>
-    <Filter Include="Pliki źródłowe\pch">
-      <UniqueIdentifier>{f3ef8480-c2bb-4636-b864-df746958df93}</UniqueIdentifier>
-    </Filter>
     <Filter Include="Pliki źródłowe\Core">
       <UniqueIdentifier>{266d80a2-a3dc-49c4-8963-68c54511c392}</UniqueIdentifier>
     </Filter>
@@ -28,6 +25,9 @@
     <Filter Include="Pliki źródłowe\Level_Control">
       <UniqueIdentifier>{3bd693c1-72f0-4c96-9b6d-114767c69eb0}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Pliki źródłowe\Engine_pch">
+      <UniqueIdentifier>{f3ef8480-c2bb-4636-b864-df746958df93}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Core\CoreMinimal.cpp">
@@ -35,12 +35,6 @@
     </ClCompile>
     <ClCompile Include="Level_control\GameController.cpp">
       <Filter>Pliki źródłowe\Level_Control</Filter>
-    </ClCompile>
-    <ClCompile Include="pch\Engine.cpp">
-      <Filter>Pliki źródłowe\pch</Filter>
-    </ClCompile>
-    <ClCompile Include="Core\CoreDefinitions.cpp">
-      <Filter>Pliki źródłowe\Core\Private</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -56,14 +50,11 @@
     <ClInclude Include="..\resources\resource.h">
       <Filter>Pliki zasobów</Filter>
     </ClInclude>
-    <ClInclude Include="Core\CoreDefinitions.h">
-      <Filter>Pliki źródłowe\Core\Public</Filter>
-    </ClInclude>
-    <ClInclude Include="pch\Engine.h">
-      <Filter>Pliki źródłowe\pch</Filter>
-    </ClInclude>
     <ClInclude Include="Level_control\GameLevel.h">
       <Filter>Pliki źródłowe\Level_Control</Filter>
+    </ClInclude>
+    <ClInclude Include="Engine_pch\Engine.h">
+      <Filter>Pliki źródłowe\Engine_pch</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
