+++ Engine/Engine_pch/Engine.h
@@ -13,3 +13,6 @@
 #else
 #define ENGINE_API _declspec(dllimport)
 #endif
+
+#define MAX_NAME_STRING 256
+#define HInstance() GetModuleHandle(NULL)
