+++ Engine/Level_control/GameController.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "GameLevel.h"
-#include "CoreMinimal.h"
 class ENGINE_API GameController {
