+++ resource.h
